+++ CHANGELOG.md
@@ -3,3 +3,17 @@
 ## [v0.0.1]
 
 - Initial release
+
+## [v0.0.13] 
+
+- FIX commands 
+
+
+## [v0.0.14] 
+
+- FIX creation of mailgun config 
+- Add get config path
+
+## [v0.0.142] 
+
+- FIX added template name for upload mailgun template